requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
frozenlist==1.4.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.1.0
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
schedule==1.2.2
setuptools==75.2.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
